elk_solution/fastapi-solution/pyproject.toml
[tool.poetry]
name = "fastapi-solution"
version = "0.1.0"
description = ""
authors = ["Ilya Klimenko <i.klimenko@lancktele.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
fastapi = "^0.111.0"
redis = "^5.0.1"
elasticsearch = {extras = ["async"], version = "^8.13.1"}
fastapi-cache2 = "^0.2.1"
pydantic-settings = "^2.2.1"
python-slugify = "^8.0.4"
orjson = "^3.10.3"
pyjwt = "^2.8.0"
sentry-sdk = "^2.11.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
pylint = "^2.17.4"
black = "^24.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
disable = "C0114,W0511, R0916, R0903"
fail-under = 9.0

[tool.pylint."MESSAGES CONTROL"]
generated-members = "df"
good-names = "df"
